requirements.txt
atlassian-python-api==3.20.1
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
Deprecated==1.2.13
idna==3.3
mysql-connector-python==8.0.28
oauthlib==3.2.0
protobuf==3.19.4
pycodestyle==2.8.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.1
toml==0.10.2
urllib3==1.26.9
wrapt==1.14.0
mysql-connector-python==8.0.28